// SampleWebApplication2021/ClientApp/src/app/services/dto.ts
// NOTE: This file was automatically generated by an automated tool.
//       Modifying the contents of this file may result in unexpected behavior.

export interface Employee {
  id: number;
  fullName: string;
  address: string;
  phoneNumber: string;
  position: EmployeePosition;
  positionID: number;
}
export interface EmployeePosition {
  id: number;
  title: string;
}